#2 / Moved collector class into consensus package.
